name changes and data type change

+++ src/main/java/gawlowska/urszula/discountApplication/repository/DiscountRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 @Repository
 public class DiscountRepository {
 
-    public static final int DEFAULT_DISCOUNT = 115;
+    public static final double DEFAULT_DISCOUNT = 115;
 
     public Discount getDiscount() {
         Discount totalDiscount = new Discount(DEFAULT_DISCOUNT);
